Enable Jenkins build on stellar and Merge PU/athena master (#49)
* From master
  * Make variables used in simd loops private
  * Fix a bug in history output
  * Fix a bug that takes up unnecessary space in restart file
  * Create CODE_OF_CONDUCT.md
  * Add Contributor Covenant badge
  * Add citation
  * Add logo
  * Update CONTRIBUTING.md
  * Remove public version release scripts
  * Add compiler configure option for nextgen Clang/LLVM Intel compiler
  * Add Zenodo
  * Fix a bug in conversion to orbital conserved variables
  * Fix calculation of energy viscous flux
  * Reorganize information given to CalcDiffusionFlux function

* From tigris
  * Add jenkins script for stellar
  * Disable single precision compile test
  * Remove redundant variable ssize_ and correct type coversion
  * Correct type for imb?v variables and use CACHELINE_BYTES definition
  * Fix a bug in OA hydro source
  * Fix a few type conversion issues
  * Changed from icpc to icpx (not for turb)
  * Change warning flags for pgen_compile
  * Suppress particle regression tests for icpx
  * Comment out remove command before running each test
  * Supperss diffusion tests

+++ src/mesh/mesh.cpp
@@ -992,7 +992,7 @@ void Mesh::OutputMeshStructure(int ndim) {
   }
   for (int i=0; i<nbtotal; i++) {
     nb_per_plevel[(loclist[i].level - root_level)]++;
-    cost_per_plevel[(loclist[i].level - root_level)] += costlist[i];
+    cost_per_plevel[(loclist[i].level - root_level)] += static_cast<int>(costlist[i]);
   }
   for (int i=root_level; i<=max_level; i++) {
     if (nb_per_plevel[i-root_level] != 0) {
@@ -1012,7 +1012,7 @@ void Mesh::OutputMeshStructure(int ndim) {
   }
   for (int i=0; i<nbtotal; i++) {
     nb_per_rank[ranklist[i]]++;
-    cost_per_rank[ranklist[i]] += costlist[i];
+    cost_per_rank[ranklist[i]] += static_cast<int>(costlist[i]);
   }
   for (int i=0; i<Globals::nranks; ++i) {
     std::cout << "  Rank = " << i << ": " << nb_per_rank[i] <<" MeshBlocks, cost = "
